영수증 기능 구현 완료 - 글자 라인 맞추기 미구현

+++ Kiosk/model.swift
@@ -52,9 +52,9 @@ class UserInfo {
 
 class Receipt: Menu {
     var count: Int
-    var priceSum: Double
+    var priceSum: Decimal
     
-    init(name: String, price: Decimal, count: Int, priceSum: Double) {
+    init(name: String, price: Decimal, count: Int, priceSum: Decimal) {
         self.count = count
         self.priceSum = priceSum
         super.init(name, price)
